Moved searchCredentials to searchService
